pyproject.toml
[project]
name = "extension-meilisearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "glootil>=0.2.2",
    "marko>=2.1.3",
    "meilisearch>=0.34.1",
    "python-frontmatter>=1.1.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "toml>=0.10.2",
    "uvicorn>=0.34.2",
]
